backend/requirements.txt
uvicornfastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
